Limit request population depth
With the "populate=deep" parameter "strapi-plugin-populate-deep" plugin,
the request selects database entries for queries it shouldn't have
access to with its token. It doesn't include anything secret in the
output, but the request takes a long time if there are a lot of
donations for a cause area. The request for the human welfare cause area
took more than 10 seconds.

Limiting the depth to 3 prevents it from getting to donations and still
gets everything we need for now. In the future we might need to think of
a better way to populate fields.