Remove proportions and separate tips
The backend doesn't really need to care about proportions, so make the
donation process a lot simpler by calculating the amounts in the
frontend (which we do anyway) and sending them in the request instead of
the proportions.

Now we can also treat tips as donations to our organization instead of
fields under donations. This makes them much nicer to deal with.